fix(olm): Decrypt the message when we create an inbound Session
The API to create an inbound Session in libolm has three distinct steps:

1. Create a Session from a pre-key message
2. Decrypt the pre-key message
3. Remove the one-time key that was used to create the Session in step 1

This API is a bit problematic since it places the burden on the user to
remove used up one-time keys. Furthermore it requires the pre-key
message to be decoded twice, once for the Session creation step and once
for the decryption step.

vodozemac removed problem one by removing the one-time key when the
Session has been created. This creates another problem, it lets users
remove private one-time keys from our pool without actually using them
up. The user can create a Session on our side without ever producing a
valid encrypted message for the Session.

This in turn results in unsuspecting users trying to communicate with us
with a one-time key that we already have removed.

Luckily we can just decrypt the message at the Session creation step and
combine the three mentioned steps into a single API call. This removes
all mentioned problems.