feat(sas): Port the calculate_mac method from libolm
The method that calculates the MAC in libolm returns an incorrectly
base64 encoded MAC. The method has been fixed upstream, but it's still
widely used.

This patch adds a method that outputs the same incorrect result, it can
be used in case compatibility with libolm is necessary.

More info can be found here:
https://gitlab.matrix.org/matrix-org/olm/-/merge_requests/16