feat!: Return the created and discarded one-time keys when generating new ones
The method to generate new one-time keys silently discards old, and
creates new one-time keys. It might be useful to the caller to know that
some one-time keys were discarded and to see which ones got newly
created.

This patch modifies the method `generate_one_time_keys()` so returns a
value showing us which, if any, one-time keys got discarded, and which
got created.