feat: Include the ratchet key in the Debug output of the Session
The ratchet key is a crucial piece of information when trying to debug a
decryption failure of a Olm Session.

The ratchet key and chain index, which are found inside a Olm message,
form a pair which uniquely identifies the individual message key that
should be used to decrypt the message.

To inspect why a decryption failure occurs it's useful to have the
ratchet key and chain index pair, this allows us to double check if the
correct chain is part of the Session.