ci: Split out the CI setups for mutation testing
This patch attempts to limit the times we run the full mutation tests.
The full tests take 1h to complete so it's not feasible to run them as
part of a pull-request CI job.

The original setup did attempt to limit this as well, but pushing
additional commits to a pull request did trigger the full run as well.